requirements.txt
atomicwrites==1.4.0
attrs==20.1.0
backports.functools-lru-cache==1.6.1
configparser==4.0.2
contextlib2==0.6.0.post1
funcsigs==1.0.2
importlib-metadata==1.7.0
more-itertools==5.0.0
numpy==1.16.6
packaging==20.4
pathlib2==2.3.5
Pillow==6.2.2
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==4.6.11
scandir==1.10.0
six==1.15.0
wcwidth==0.2.5
zipp==1.2.0
